fdat-py/requirements.txt
pywebview
lxml
pyinstaller